+++ src/Infrastructure/Queue/EnqueuerContainer.php
@@ -22,6 +22,8 @@ final class EnqueuerContainer implements IEnqueuerContainer
     private array $bindings = [];
 
     /**
+     * EnqueuerContainer constructor.
+     *
      * @param Closure(): object $default
      */
     public function __construct(private readonly Closure $default)